requirement.txt
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asttokens==2.4.1
backcall==0.2.0
bash==0.6
Cartopy @ file:///D:/MyOfficeFiles/2%E7%A7%91%E7%A0%94/8.%E5%90%88%E4%BD%9C%E7%A0%94%E7%A9%B6/%E5%BB%B6%E5%AE%89%E5%B8%82%E6%B0%94%E8%B1%A1%E5%B1%80/Codes/cartopy_setup/Cartopy-0.20.2-cp38-cp38-win_amd64.whl#sha256=feae60ea2450e50a14caae345e94b56d4bce0bb65ac7977c80b144398abb58b5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
einops==0.6.0
executing==2.0.1
fonttools==4.46.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
jmespath==0.10.0
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcls==0.23.0
mmcv-full==1.5.2
mmdet==2.24.1
# Editable install with no version control (mmfewshot==0.1.0)
-e g:\zhailonglong\mobilevit\mmfewshot-main
model-index==0.1.11
nest-asyncio==1.5.8
numpy==1.24.4
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.32
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycocotools==2.0.7
pycocotools-windows==2.0.0.2
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
pyparsing==3.1.1
pyproj @ file:///D:/MyOfficeFiles/2%E7%A7%91%E7%A0%94/8.%E5%90%88%E4%BD%9C%E7%A0%94%E7%A9%B6/%E5%BB%B6%E5%AE%89%E5%B8%82%E6%B0%94%E8%B1%A1%E5%B1%80/Codes/cartopy_setup/pyproj-3.2.1-cp38-cp38-win_amd64.whl#sha256=50d312cb7610f93f02f07b7da5b96469c52645717bebe6530ac7214cc69c068e
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.28.2
rich==13.4.2
scikit-learn==1.3.2
scipy==1.10.1
Shapely @ file:///D:/MyOfficeFiles/2%E7%A7%91%E7%A0%94/8.%E5%90%88%E4%BD%9C%E7%A0%94%E7%A9%B6/%E5%BB%B6%E5%AE%89%E5%B8%82%E6%B0%94%E8%B1%A1%E5%B1%80/Codes/cartopy_setup/Shapely-1.8.2-cp38-cp38-win_amd64.whl#sha256=27af42a2f0374c8d1c0f62a8272ac28816c8e60f642e2934f46d0015e7b1c347
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
terminaltables==3.1.10
threadpoolctl==3.3.0
tomli==2.0.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.3.3
tqdm==4.65.2
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.9
yapf==0.40.1
zipp==3.17.0
